Add support for Android 10.
Replacing libsecure_storage.so with an older version no longer works on
Android 10.

Instead, we create a patched version of libbluetooth.so at installation
time from the version found on the device and use that to mask the
original.

Tested and verified on S10 series BSKO firmware.

Thank you to Arthur Trouillot for finding the byte sequence to patch.